feat(ui): add history and diff views with mocked data
